pyproject.toml
[project]
        name = "SweepGraph"
        version = "1"
        description = "Text mining sweeps to build a Neo4j-ready knowledge graph using Gemini 2.5 Pro or any LLM"
        authors = [
            { name = "Magnús Smári Smárson" }
        ]
        requires-python = ">=3.11"
        dependencies = [
            "neo4j>=5.0.0",
            "google-generativeai>=0.3.0",
            "python-dotenv>=1.0.0",
            "rich>=13.0.0",
            "anthropic>=0.25.0",
            "streamlit>=1.37",
            "streamlit-agraph>=0.0.45",
            "plotly>=5.22",
            "pandas>=2.2",
        ]

        [project.urls]
        Homepage = "https://www.smarason.is"

        [project.optional-dependencies]
        dev = [
            "pytest>=7.0.0",
            "black>=23.0.0",
            "ruff>=0.1.0",
        ]

        [tool.uv]
        dev-dependencies = []

        [build-system]
        requires = ["hatchling"]
        build-backend = "hatchling.build"
        